Uncommented code for foreach support with k8s
KubernetesPodOperator version 4.2.0 renamed `resources` to
`container_resources`
- Check : (https://github.com/apache/airflow/pull/24673) /
- (https://github.com/apache/airflow/commit/45f4290712f5f779e57034f81dbaab5d77d5de85)

This was done because `KubernetesPodOperator` didn't play nice with dynamic task mapping and they had to deprecate the `resources` argument. Hence the below codepath checks for the version of `KubernetesPodOperator`
and then sets the argument. If the version < 4.2.0 then we set the argument as `resources`.
If it is > 4.2.0 then we set the argument as `container_resources`
The `resources` argument of KuberentesPodOperator is going to be deprecated soon in the future.
So we will only use it for `KuberentesPodOperator` version < 4.2.0
The `resources` argument will also not work for foreach's.